Presentations/ffmpeg-python/requirements.txt
ffmpeg-python
graphviz
numpy
tensorflow<2